modules/sniper/main.py
import asyncio
import os
import logging
from telebot.async_telebot import AsyncTeleBot

# Логирование
logging.basicConfig(level=logging.INFO, filename=os.path.join(os.path.dirname(__file__), '../../logs/app.log'), format='%(asctime)s - %(levelname)s - %(message)s')

BOT_TOKEN = os.getenv("BOT_TOKEN", "8780387143:AAEovV_r-tD8oggABGgUo-f93jduMz0r78g")
ADMIN_ID = int(os.getenv("ADMIN_ID", "8335925220"))
bot = AsyncTeleBot(BOT_TOKEN)

async def check_new_orders():
    # Здесь будет API Freelancehunt / Парсер
    logging.info("Снайпер: проверка новых заказов...")
    return []

async def sniper_loop():
    await bot.send_message(ADMIN_ID, "🎯 <b>Снайпер запущен.</b> Мониторинг заказов активирован (Интервал: 60 сек).", parse_mode="HTML")
    while True:
        try:
            orders = await check_new_orders()
            if orders:
                pass # Отправка инлайн-кнопок с заказами
            await asyncio.sleep(60)
        except Exception as e:
            logging.error(f"Sniper fault: {e}")
            await asyncio.sleep(5)

if __name__ == "__main__":
    asyncio.run(sniper_loop())
