projects/sas_monitor_bot/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()
BOT_TOKEN = os.getenv("BOT_TOKEN")
ADMIN_ID = int(os.getenv("ADMIN_ID")) if os.getenv("ADMIN_ID") else 0
N8N_WEBHOOK_URL="http://твой-ip:5678/webhook/sas-report"
