requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
pyTelegramBotAPI==4.15.4
